Connector: Workaround to duplicatekey errors
The multiple DB loads as until loadLovefieldDB() returned, db was still
null, causing it to be loaded again if any other requests for the DB
were made during this time.

I think this was part of the cause of the duplicate key errors, such as:
Constraint error: (201) Duplicate keys are not allowed, index: Key.pkKey, key: 3

This could result in very bad issues when we also had multiple syncs at
the same time from the connector.