Connector: P2S input fix
We must use the inputs given to us as boxes to sign as even if we arent
directly signing them (P2S inputs) they need to be included in the tx
signing in sigma-rust. P2S inputs weren't found in the user's UTXOs so
they were silently filtered from the signing process unintentionally.

We will still need to update this later to get the keys in the same way
somehow or find a different way for the cases where it's a UTXO we'd
have control to sign over but it hasn't yet been put in our DB (e.g.
pending txs which was the reason why we changed UnsignedInputs to
include the full data in the first place)