Ergo Connector: Fix crash when signing a TX that mints new assets
We rely on the ergo explorer backend to look up token info, but this
won't work when we're minting a new token.

So instead, we will check if the tx minted a token and create the token
info from the registers according to EIP-0004

This PR does not read all the info as we'll need to update sigma-rust to
decode values first to be able to decode the values I think, so they're
all mocked out as the "TODO" token for now.